[feat]: 스크롤에 따른 게시글 제목 표시 헤더 구현
- 헤더 높이만큼 스크롤 됐을 시, 제목 텍스트로 전환하여 보여주는 기능 추가
- 전환 시 부드러운 애니메이션 적용
